+++ src/mappings/mappingHandlers.ts
@@ -147,33 +147,6 @@ export async function handleCw20BalanceTransfer(event: CosmosEvent): Promise<voi
   await saveCw20BalanceEvent(`${id}-debit`, fromAddress, BigInt(0)-BigInt(amount), contract, event);
 }
 
-export async function handleGovProposalVote(event: CosmosEvent): Promise<void> {
-  const msg: CosmosMessage<GovProposalVoteMsg> = event.msg;
-  logger.info(`[handleGovProposalVote] (tx ${msg.tx.hash}): indexing GovProposalVote ${messageId(msg)}`)
-  logger.debug(`[handleGovProposalVote] (event.msg.msg): ${JSON.stringify(msg.msg, null, 2)}`)
-
-  const id = messageId(msg);
-  const option = msg?.msg?.decodedMsg?.option;
-  const proposalId = msg?.msg?.decodedMsg?.proposalId, voter = msg?.msg?.decodedMsg?.voter;
-
-  if (!option || !proposalId || !voter) {
-    logger.warn(`[handleGovProposalVote] (tx ${event.tx.hash}): cannot index event (event.event): ${JSON.stringify(event.event, null, 2)}`)
-    return
-  }
-
-  const vote = GovProposalVote.create({
-    id,
-    proposalId: proposalId,
-    voterAddress: voter,
-    option: Object.values(GovProposalVoteOption)[option],
-    messageId: id,
-    transactionId: msg.tx.hash,
-    blockId: msg.block.block.id,
-  });
-
-  await vote.save();
-}
-
 export async function handleLegacyBridgeSwap(event: CosmosEvent): Promise<void> {
   const msg: CosmosMessage<LegacyBridgeSwapMsg> = event.msg
   const id = messageId(msg);
@@ -265,3 +238,4 @@ async function saveCw20BalanceEvent(id: string, address: string, amount: BigInt,
 export * from "./primitives";
 export * from "./bank";
 export * from "./dist";
+export * from "./gov";
